Split off the VerifyNVQIRCallOps pass. (#1569)
Remove Hungarian notation.
