+++ lib/Optimizer/CodeGen/CMakeLists.txt
@@ -20,6 +20,7 @@ add_cudaq_library(OptCodeGen
   RemoveMeasurements.cpp
   TranslateToIQMJson.cpp
   TranslateToOpenQASM.cpp
+  VerifyNVQIRCalls.cpp
 
   DEPENDS
     CCDialect
